Add methods to restore user session from cache without showing sign in UI
In 1.0.4, support was added to cache and restore a session so that it
was unnecessary to prompt the user to sign in each time the service was
accessed. However, this new method coupled together restoring the
session and showing the sign in UI if the credentials were invalid.

This is problematic because there are several scenarios where it's
important to separate out restoring a session and showing sign in UI:
- Restoring a session from a background task where UI cannot be raised.
- Restoring a session from within an app, but where it's important for
the app to have full control over the sign in flow. For example, if an
app automatically contacts the service on app launch, it would be far
more elegant for the app to show an inline message that the user must
re-authenticate, rather than just randomly showing the sign in UI
without explanation.

To address these scenarios, this change adds a new set of methods that
only restores the session without showing the sign in UI.